+++ Package.swift
@@ -23,7 +23,6 @@ let package = Package(
         .package(url: "https://github.com/danielsaidi/OnboardingKit.git", .upToNextMajor(from: "7.0.2")),
         .package(url: "https://github.com/danielsaidi/StoreKitPlus.git", .upToNextMajor(from: "0.6.0")),
         .package(url: "https://github.com/danielsaidi/SwiftUIKit.git", .upToNextMajor(from: "5.1.1")),
-        .package(url: "https://github.com/danielsaidi/SystemNotification.git", .upToNextMajor(from: "1.2.1"))
     ],
     targets: [
         .target(
@@ -33,8 +32,7 @@ let package = Package(
                 "ConfettiSwiftUI",
                 "OnboardingKit",
                 "SwiftUIKit",
-                "StoreKitPlus",
-                "SystemNotification"
+                "StoreKitPlus"
             ],
             resources: [.process("Resources")]
         ),