+++ Package.swift
@@ -18,10 +18,10 @@ let package = Package(
         ),
     ],
     dependencies: [
-        .package(url: "https://github.com/danielsaidi/BadgeIcon.git", .upToNextMajor(from: "0.6.0")),
+        .package(url: "https://github.com/danielsaidi/BadgeIcon.git", .upToNextMinor(from: "0.6.0")),
         .package(url: "https://github.com/simibac/ConfettiSwiftUI.git", .upToNextMajor(from: "1.1.0")),
         .package(url: "https://github.com/danielsaidi/OnboardingKit.git", .upToNextMajor(from: "7.0.2")),
-        .package(url: "https://github.com/danielsaidi/StoreKitPlus.git", .upToNextMajor(from: "0.6.0")),
+        .package(url: "https://github.com/danielsaidi/StoreKitPlus.git", .upToNextMinor(from: "0.6.0")),
         .package(url: "https://github.com/danielsaidi/SwiftUIKit.git", .upToNextMajor(from: "5.1.1")),
     ],
     targets: [